Reorganiza estructura de carpetas del proyecto
