+++ src/main.jsx
@@ -1,6 +1,6 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import App from './App.jsx'
+import App from './components/App.jsx'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
